contracts-stylus: reorganize module structure
