+++ contracts-stylus/src/contracts/test_contracts/dummy_erc20.rs
