+++ contracts-stylus/src/lib.rs
@@ -1,23 +1,8 @@
 #![no_main]
 #![no_std]
 
-mod constants;
-mod interfaces;
+mod contracts;
 mod utils;
-
-#[cfg(any(feature = "darkpool", feature = "darkpool-test-contract"))]
-mod darkpool;
-
-#[cfg(feature = "verifier")]
-mod verifier;
-
-#[cfg(any(
-    feature = "precompile-test-contract",
-    feature = "verifier-test-contract",
-    feature = "darkpool-test-contract",
-    feature = "dummy-erc20"
-))]
-mod test_contracts;
 
 extern crate alloc;
