+++ contracts-stylus/src/utils/constants.rs
