+++ contracts-stylus/src/utils/interfaces.rs
