+++ contracts-stylus/src/utils/mod.rs
@@ -1,0 +1,7 @@
+//! Various utilities used throughout the contracts,
+//! including constants, ABI interfaces, and backend definitions
+
+pub mod backends;
+pub mod constants;
+pub mod helpers;
+pub mod interfaces;
